finaliza css de homePage
