+++ frontend/src/components/cardLoja/index.css
@@ -1,7 +1,38 @@
-.cardWraper{
+/* .cardWraper{
     display: flex;
     flex-direction: column;
     border-bottom: 1px solid rgba(197, 197, 197, 0.5);
     align-items: center;
     gap: 0.5rem;
+} */
+
+.cardWraper{
+    display: flex;
+    flex-direction: row;
+    border-bottom: 1px solid #969696;
+    align-items: center;
+    width: 100%;
+    height: 55px;
+    text-align: start;
+    justify-content: space-between;
+    padding: 3px;
+}
+
+.cardWraper:hover{
+    border: 1px solid #969696;
+    color: var(--color-secondary);
+}
+
+.icons {
+    padding: 3px;
+    margin-inline: 2px;
+    cursor: pointer;
+}
+
+.edit-icon:hover{
+    color: var(--color-secondary);
+}
+
+.delete-icon:hover{
+    color: var(--color-secondary);
 }
