+++ frontend/src/components/cardLoja/index.tsx
@@ -1,5 +1,6 @@
 import './index.css'
-
+import editIcon from '../../images/edit-icon.svg'
+import deleteIcon from '../../images/delete-icon.svg'
 
 interface CardProps {
     nome?: String;
@@ -10,18 +11,29 @@ interface CardProps {
 
 }
 
+export function CardLoja({ nome, cnpj, telefone, cep, bairro }: CardProps) {
+    return (
+        <div className="cardWraper">
+            <text className='loja-nome' >{nome}</text>
+            <div className='icons'>
+                <img className='edit-icon' onClick={() => alert("editar")} src={editIcon} />
+                <img className='delete-icon' onClick={() => alert("deletar")} src={deleteIcon} />
+            </div>
+        </div>
+    );
+}
 
-export function CardLoja({nome, cnpj, telefone, cep, bairro}: CardProps){
+/* export function CardLoja({nome, cnpj, telefone, cep, bairro}: CardProps){
     return(
         <div className="cardWraper">
                 <p>Nome: {nome}</p>
                 <p>CNPJ:  {cnpj}</p>
                 <p>Telefone Primário:  {telefone}</p>
                 <p>CEP:  {cep}</p>
-                <p>Bairro:  {bairro}</p>  
+                <p>Bairro:  {bairro}</p>
         </div>
     );
-}
+} */
 
 // export function CardLoja(nome: String, cnpj:String, telefone: String, telefone2: String, cep: String, endereco: String){
 //     return(
