+++ frontend/src/components/header/index.tsx
@@ -1,8 +1,38 @@
 import './index.css'
 import CapitalLogo from '../../images/capitalLogo.svg'
+import userIcon from '../../images/user-icon.svg'
+import { useContext, useEffect, useState } from 'react'
+import AuthContext from '../../contexts/auth'
+import { useNavigate } from 'react-router-dom'
 
+interface User {
+    name: string,
+    email: string,
+    password: string,
+    id: Number
+}
 
 export function Header(){
+    const [user, setUser] = useState<User>();
+    const context = useContext(AuthContext);
+    let navigate = useNavigate();
+
+    useEffect(() => {
+        const storagedUser = localStorage.getItem('@App:user');
+        if (storagedUser) {
+            setUser(JSON.parse(storagedUser));
+        }
+    }, []);
+
+    function handleClick() {
+        try {
+            context.logOut();
+            navigate("/");
+        } catch (error) {
+            alert("Não foi possivel realizar a açao")
+        }
+    }
+
     return (
         <div className="header">
 
@@ -12,9 +42,9 @@ export function Header(){
                     <span>Captação de Clientes</span>
                 </div>    
 
-                <div className="user">
-                    <img src= ''/>
-                    <span>Lorem ipsum.</span>
+                <div className="user" onClick={() => handleClick()} >
+                    <img className="user-icon" src={userIcon} />
+                    <span>{user?.name}</span>
                 </div>
             </div>
