+++ frontend/src/pages/homeAdmin/index.css
@@ -1,9 +1,68 @@
-#mainDiv {
-    background-color: #F6F6F6;
-    width: 100vw;
-    height: 100vh;
+.main {
+    display: flex;
+    flex-direction: column;
+    background-color: var(--background);
+    font-family: var(--font-family);
+}
+
+.sidebar-index {
+    display: flex;
+}
+
+.index {
+    width: 100%;
+    height: 100%;
+    text-align: center;
+}
+
+.header-lojas {
+    width: 100%;
+    height: 6rem;
     display: flex;
     flex-direction: row;
-    justify-content: center;
     align-items: center;
+    justify-content: space-around;
+    /* border-bottom: 1px solid rgba(217, 217, 217, 0.5); */
+}
+
+#totalLojas {
+    color: #969696;
+}
+
+#adicionarLojas {
+    border: 1px solid #FF834F;
+    border-radius: 4px;
+    width: 140px;
+    height: 40px;
+    box-sizing: border-box;
+    color: var(--color-primary);
+    text-align: center;
+}
+
+#addLojasText {
+    margin-top: 11px;
+}
+
+#titulo {
+    width: 35px;
+    height: 18px;
+    /* margin-left: 41%; */
+    font-size: 20px;
+    justify-content: center;
+    text-align: center;
+    color: #000;
+}
+
+.header-lojas>span {
+    color: black;
+}
+
+.cards {
+    height: calc(100vh - 16rem);
+    overflow-y: scroll;
+    /* display: grid; */
+    gap: 1rem;
+    width: 80%;
+    margin-top: 30px;
+    margin-left: 10%;
 }
