+++ frontend/src/pages/login/index.css
@@ -3,7 +3,6 @@
     width: 100vw;
     height: 100vh;
     font-family: var(--font-family);
-    position: relative
 }
 
 #rodape {
@@ -15,6 +14,7 @@
     color: #fff;
     padding: 10px;
     display:inline-flex;
+    background-color: var(--color-primary);
 }
 
 #body {
