+++ frontend/src/pages/login/index.tsx
@@ -29,13 +29,14 @@ const Login = () => {
             <div id="rodape" >
                 <img
                     src={CapitalLogo}
-                    style={{ height: 36, width: 117.83, marginRight: 30}}
+                    style={{ height: 36, width: 117.83, marginRight: 30 }}
                     alt="Capital atacadista"
                 />
-                <text style={{marginTop: 10}} >Capitação de Clientes</text>
+                <text style={{ marginTop: 10 }} >Capitação de Clientes</text>
             </div>
             <div id="body">
                 <div id="box" >
+                    <br /><br />
                     <h2>Login</h2>
                     <div id="formDiv">
                         <div className="inputDiv" >
