--- frontend/src/pages/lojas/index.css
@@ -1,38 +1,0 @@
-.main{
-    display: flex;
-    flex-direction: column;
-}
-
-.sidebar-index{
-    display: flex;
-}
-
-.index{
-   width: 100%; 
-}
-
-.header-lojas{
-    width: 100%;
-    height: 6rem;
-
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: space-around;
-
-    border-bottom: 1px solid rgba(217,217,217,0.5);
-    
-}
-
-.header-lojas>span{
-    color: black;
-}
-
-.cards{
-    height: calc(100vh - 11rem);
-    overflow-y: scroll;
-    display: grid;
-    gap: 1rem;
-}
-  
-
