+++ frontend/src/services/api.ts
@@ -1,7 +1,7 @@
 import axios from "axios";
 
 const api = axios.create({
-  baseURL: process.env.REACT_APP_URL_API,
+  baseURL: "http://localhost:4000",
 });
 
 export default api;
